Fix references to the renamed file.
